+++ docs/diary/2026-04-24-reflection-fr-276-prompt-caching.md
