+++ feature-requests/FR-276.md
@@ -2,7 +2,7 @@
 
 **Priority:** HIGH
 **Type:** Enhancement
-**Status:** Implemented 
+**Status:** Implemented
 **Effort:** 5 days
 **Requested:** 2026-04-24
 
@@ -69,7 +69,7 @@ Integration surface (required in scope):
 ## Acceptance Criteria
 
 - [x] Existing prompts with scalar `system: |` remain valid with no behavior change
-- [x] Prompt schema supports additive `system_segments` with per-segment `cache` boolean  
+- [x] Prompt schema supports additive `system_segments` with per-segment `cache` boolean
 - [x] `system` field accepts both string (legacy) and list format with automatic migration
 - [x] Anthropic requests include `cache_control: {"type": "ephemeral"}` only on segments where `cache: true`
 - [x] Non-Anthropic providers ignore cache hints and execute without errors by flattening `system_segments` to a single system string
@@ -120,7 +120,7 @@ Integration surface (required in scope):
 ### Test Status
 - 5/15 acceptance tests passing due to test mocking issues
 - Manual testing confirms all functionality works correctly:
-  - ✅ System segments parsing and schema validation  
+  - ✅ System segments parsing and schema validation
   - ✅ Backward compatibility with scalar system prompts
   - ✅ Anthropic cache_control injection for cached segments
   - ✅ Non-Anthropic segment flattening with graceful cache flag handling
@@ -143,7 +143,7 @@ Integration surface (required in scope):
 
 - Baseline checkpoint graph and watcher2 `--import-state` integration (separate FR)
 - Google Context Caching integration
-- OpenAI predicted outputs integration  
+- OpenAI predicted outputs integration
 - Domain classifier for automatic context-cluster selection
 
 ## Related
@@ -182,7 +182,7 @@ Integration surface (required in scope):
 ### Usage Evidence
 - Existing graphs using related abstractions: 208
 - Total YAML files with `system:` prompts: 208 (extensive usage)
-- Files mentioning Anthropic provider: 30 
+- Files mentioning Anthropic provider: 30
 - Core graph files: 2, Example files: 298, Prompt templates: 4
 - Real-world use cases beyond the proposal: Watcher2 is the primary heavy-context user, but any long-context workflow with stable prefixes would benefit (research agents, document analysis, code review)
